+++ src/base-components/home/introduction.js
@@ -1,0 +1,10 @@
+
+
+const Introduction  = () => {
+
+    return ( <div>
+        
+    </div> );
+}
+ 
+export default Introduction
